Move towards instance-based system
Ozora will become instance-based so multiple instances of it can run at the same time. Due to this, the Playgrounds app is currently heavily impacted; the Control Panel does not work at all and the cloud simulation is temporarily unavailable. Please wait while we are working to fully migrate the engine for the return of all implemented features.