Added trailing system
Now, the mouse cursor can be disengaged from the object (in case you only want to drag the sun around or the mouse pointer leaves the window). It will keep its momentum and slow down using an adjustable drag coefficient. Obviously, this is the most fun if you enable the collision feature and enable bounces as well as set the rubber banding force to high to allow users to throw an object against a window border and have it bounce around the window.

Theoretically, you can now also set the drag coefficient to 0 and the bounce kinetic energy retention factor to 1 to get limitless motion.