Started working on ParulAI
ParulAI will be able to control a bird sprite's animations and movement to add a lively touch to a nature scene. Work in progress, not usable.

For simplicity reasons, the Ozora Playgrounds app is currently limited to the birds simulation, which is not functional yet. Use an older build to try out the bouncy simulation.