Fixed flickering on sprite transition
When the sprites transitioned, there used to be a flickering effect due to a bug with the Image control in WinUI3. This is now mitigated by placing another version on top of the old version, refreshing the old version, and once it's done, removing the temporary version to only show the now new version. It's a bit of a hacky workaround, but what are you gonna do about it?