Made settings neighboring resting spots easier
When setting a spot as a spot to a current spot's right, the current spot will be set as to the left of that spot automatically. This is intended to make positioning much easier if you intend on positioning batches of three spots each because you will only have to set the middle spot's neighbors, the rest will adapt automatically. Longer chains will also be simpler to set up with this.