Fixed ParulAI crash
This crash occurred as the bird was always instructed to sit down if it reached its target and chat with its neighbors. Now, because the bird was unable to talk to its neighbors due to having no neighbors, the RestingSpot would return null and thus, the algorithm would throw an exception.